largimi i db ne admin branch

+++ admin/delete_course.php
@@ -1,15 +1,14 @@
 <?php
 session_start();
-include "../db.php";
+include "data/data.php";
 include "classes/Admin.php";
 
-// if ($_SESSION['is_admin'] != 1) {
-//     die("Access denied");
-// }
+if ($_SESSION['user']['role'] != "admin") {
+    die("Access denied");
+}
 
-$admin = new Admin($conn);
+$admin = new Admin($users, $courses);
 
 $admin->deleteCourse($_GET['id']);
 
 header("Location: admin.php");
-?>
